requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
cmake==3.31.0.1
easydict==1.13
filelock==3.16.1
ftfy==6.3.1
graphviz==0.8.4
grpcio==1.67.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
lit==18.1.8
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
mxnet==1.9.1
networkx==3.2.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-clip==1.0.1
opencv-python==4.5.5.64
packaging==24.2
pillow==11.0.0
protobuf==5.28.3
regex==2024.11.6
requests==2.32.3
scikit-learn==1.1.1
scipy==1.13.1
six==1.16.0
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.0
triton==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
